requirements.txt
flask
onnxruntime
pillow
fpdf2
python-dotenv
gunicorn